@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Text;
 using System.Text.Json;
+using System.Threading.Tasks;
 using static BBDown.BBDownLogger;
 using static BBDown.BBDownUtil;
 
@@ -10,16 +11,16 @@ namespace BBDown
 {
     class BBDownSpaceVideoFetcher : IFetcher
     {
-        public BBDownVInfo Fetch(string id)
+        public async Task<BBDownVInfo> FetchAsync(string id)
         {
             id = id.Substring(4);
             string userInfoApi = $"https://api.bilibili.com/x/space/acc/info?mid={id}&jsonp=jsonp";
-            string userName = GetValidFileName(JsonDocument.Parse(GetWebSource(userInfoApi)).RootElement.GetProperty("data").GetProperty("name").ToString());
+            string userName = GetValidFileName(JsonDocument.Parse(await GetWebSourceAsync(userInfoApi)).RootElement.GetProperty("data").GetProperty("name").ToString());
             List<string> urls = new List<string>();
             int pageSize = 100;
             int pageNumber = 1;
             string api = $"https://api.bilibili.com/x/space/arc/search?mid={id}&ps={pageSize}&tid=0&pn={pageNumber}&keyword=&order=pubdate&jsonp=jsonp";
-            string json = GetWebSource(api);
+            string json = await GetWebSourceAsync(api);
             var infoJson = JsonDocument.Parse(json);
             var pages = infoJson.RootElement.GetProperty("data").GetProperty("list").GetProperty("vlist").EnumerateArray();
             foreach (var page in pages)
@@ -31,7 +32,7 @@ namespace BBDown
             while (pageNumber < totalPage)
             {
                 pageNumber++;
-                urls.AddRange(GetVideosByPage(pageNumber, pageSize,  id));
+                urls.AddRange(await GetVideosByPageAsync(pageNumber, pageSize,  id));
             }
             File.WriteAllText($"{userName}的投稿视频.txt", string.Join('\n', urls));
             Log("目前下载器不支持下载用户的全部投稿视频，不过程序已经获取到了该用户的全部投稿视频地址，你可以自行使用批处理脚本等手段调用本程序进行批量下载。如在Windows系统你可以使用如下代码：");
@@ -43,11 +44,11 @@ pause");
             throw new Exception("暂不支持该功能");
         }
 
-        List<string> GetVideosByPage(int pageNumber, int pageSize, string mid)
+        async Task<List<string>> GetVideosByPageAsync(int pageNumber, int pageSize, string mid)
         {
             List<string> urls = new List<string>();
             string api = $"https://api.bilibili.com/x/space/arc/search?mid={mid}&ps={pageSize}&tid=0&pn={pageNumber}&keyword=&order=pubdate&jsonp=jsonp";
-            string json = GetWebSource(api);
+            string json = await GetWebSourceAsync(api);
             var infoJson = JsonDocument.Parse(json);
             var pages = infoJson.RootElement.GetProperty("data").GetProperty("list").GetProperty("vlist").EnumerateArray();
             foreach (var page in pages)
