@@ -1,11 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
+using System.Threading.Tasks;
 
 namespace BBDown
 {
     interface IFetcher
     {
-        BBDownVInfo Fetch(string id);
+        Task<BBDownVInfo> FetchAsync(string id);
     }
 }
